refactor: split LeaderboardClient into focused sub-components
